April's Main Menu with parallax and music
I implemented a working main menu with volume control, credits, and an easy to change script for the scene manager. As such, one needs to go into the MainMenu script and set the proper scene needed to start the game.

Moreover, the parallax script works for the background perfectly as intended, and I wrote it such that it works on both the x and z axes. All that needs to be done is to change the parallax effect parameters for each layer of sprites/objects that need to be under the parallax effect.

Lastly, I cleaned up the main theme so we have a nice spacey jam. Also, I used 'Worldlines' as a working title, which is just paths in spacetime. Feel free to change it, it's there as a placeholder.