Multiple changes
First, I added comments to both parallax scripts. The Main Menu one works on a different plane than the Objects one, so they must stay separate. I also renamed each one according to Soph's comments on my previous PR.

Next, I added music to the main gameplay, parallax, and an audio button. Note that I need to add a script that prevents movement when the player clicks here, but it works to lower the volume as intended.

Finally, I adjusted the lobby UI to include parallax, music, and an appropiate font with a nice colorblind friendly palette (the same one from the MM).